Add exception handling around initial create/user
Unfortuantely when configuring an ME4 array for the first time, the
first thing that must be done is to create an initial admin user.
Unfortunately the api returns plain text response to this first call,
which caused our previous logic to thrown an exception as it wasn't
valid JSON as expected.

This commit adds additional error handling to handle this scenario