Add retry adapter + backoff to session requests
This changes includes one minor fix and a new feature.

Minor fix: Fixed the _login function to utilise the requests.session
(this means that the the initial GET call to the login api endpoint
benefits from retries/timeouts etc, like other calls).

Feature: Added a requests transport adapter to implement retries on
certain HTTP requests. This means that failed calls (such as
ConnectionError) get retried in cases where the server fails
prematurely.

This is common in certain scenarios with the ME4 array, such as when
changing network settings on the controllers (and thus need to establish
new Session)